cli tool implementation
