+++ pyproject.toml
@@ -1,0 +1,23 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "fastbff"
+version = "0.1.0"
+description = "Config-driven backend for frontend"
+readme = "README.md"
+authors = [{ name = "Your Name", email = "you@example.com" }]
+dependencies = [
+    "fastapi",
+    "uvicorn",
+    "pyyaml",
+    "typer[all]"
+]
+requires-python = ">=3.8"
+
+[project.scripts]
+fastbff = "fastbff.cli:app"
+
+[tool.setuptools.packages.find]
+where = ["."]
